# PruebaTecnicaInventario.API/Controllers/ClientController.cs
using Microsoft.AspNetCore.Mvc;
using PruebaTecnicaInventario.Core.Interfaces;
using PruebaTecnicaInventario.Core.Models;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Net;
using System.Threading.Tasks;

namespace PruebaTecnicaInventario.API.Controllers
{
    [Route("api/[controller]")]
    [ApiController]
    public class ClientController : Controller
    {
        private readonly IClientService _clientService;

        public ClientController(IClientService clientService)
        {
            _clientService = clientService;
        }

        [HttpPost]
        public async Task<IActionResult> Post(Client client)
        {
            await _clientService.InsertClient(client);
            return Ok(client);
        }

        [ProducesResponseType((int)HttpStatusCode.OK)]
        [ProducesResponseType((int)HttpStatusCode.BadRequest)]
        [HttpGet("{id}")]
        public async Task<IActionResult> GetClient(int id)
        {
            var client = await _clientService.GetClient(id);
            return Ok(client);
        }

        [ProducesResponseType((int)HttpStatusCode.OK)]
        [ProducesResponseType((int)HttpStatusCode.BadRequest)]
        [HttpGet]
        public IActionResult GetClients()
        {
            var client =  _clientService.GetClients();
            return Ok(client);
        }

        [HttpPut]
        public async Task<IActionResult> Put(Client client)
        {
            var response = await _clientService.UpdateClient(client);
            return Ok(response);
        }

        [HttpDelete]
        public async Task<IActionResult> Delete(int id)
        {
            var response = await _clientService.DeleteClient(id);
            return Ok(response);
        }
    }
}